- Removed all const qualifiers (except const char*): they often   required dangerous casts to remove the qualifier.
  Some (not useful anymore) casts may be still there.

- Some bugs correction (in mlapronidl, in particular)

- Improved/Modified Makefiles:

  * variable HAS_OCAML introduced in Makefile.config.model

  * Most variables factorized in Makefile.config: (CC, OCAMLC, CFLAGS, ...)

  * For OCaml .idl files, they do not generate automatically .c, .ml, .mli files.
    Target rebuild should be explictly provided.



git-svn-id: svn://scm.gforge.inria.fr/svnroot/apron/trunk@242 d6d8869d-df28-4c9d-b3d8-ddc8926f8a8a

+++ newpolka/pk_config.h
@@ -111,8 +111,8 @@ static inline size_t numint_size2(numint_t a)
 
 
 /* Do not change ! */
-static const int polka_cst = 1;
-static const int polka_eps = 2;
+static const size_t polka_cst = 1;
+static const size_t polka_eps = 2;
 
 #ifdef __cplusplus
 }
